Fix multi stream compression bugs causing invalid zlib headers
- We were using || as a delimiter which caused issues when | showed up anywhere in data
to be compressed. Fix was to use unique bytes in delimiter and check that delimiter does
not occur in data
- If compressed data contained extra occurrences of the delimiter the decompressor would
split a single stream causing header checks to fail. Fix was to use a single byte as
delimiter (different than first delimiter) and escape this byte into a different two byte
sequence when it occurs in compressed data